ramping optimize
some small bugs fixed. Now, if the move is less than the step size in x, y, and z, it just sets the voltage and doesn't try to set up a sequence to send a single voltage. I was running into issue with the server where it would try to write the task before starting it. This fixed it