rabi working
Both an issue of a train i nthe pusle sequence for the apd tagger (0, LOW), which when set to the extra time from the pi pulse (tau_max - tau), fixed the issue of reference counts being too low.
Also, I updated the laser timing, with new sequence where reference pusle is same length as gate, the timing for the laser needs to be fairly accurate.