pesr fix
In the pesr code, we pass the pi pulse time as the "tau" and as "max tau", meaning that the end of the sequence has 0 extra time. The gate pulse is on up until the very end. For some reason, on rabi_pc, that causes all but the first point to be collected ,and all other counts to be 0. If I add just 10 ns onto the end time for all the channels for the sequence, it fixes this issue, and the measurement works. I am not sure why this doesn't cause an issue on hahn_pc...