determine charge readout params
Updating this routine to estimate the threshold value for single shot charge state readout. It offers both a quick and dirty way and the more in depth way using a photon statistics model.

One issue with the full model: I need the number of counts from the experiment to be in a single list.