cleaning up opx code
I cleaned up the opx code to make it easier to read and work with using the wiki page. I cleaned up a lot of the sequence files. It shouldn't affect them, but if you are noticing issues, go back to before this commit and test it out with the previous file which we know worked