cleaning up opx code more
cleaning up opx code more. also added back in a toolbelt function for getting the uwave pulse info for the opx sequences and added back in part of the toolbelt function that gets the laser pulse info for the opx sequences. For digital pulses, if that power is 0 it makes it LOW. otherwise it is high